[TS-426] Refactor: 별자리 메인 이미지 PNG -> WEBP로 변경
